Two methods were removed from BuildIndex: (1) GetLuceneIndex (2) GetFieldsToIndex
These were required for the previous implementation of IncrementalDocVectors, but the revised implementation does not require these, as the Lucene Index and FieldsToIndex String[] are passed to the constructor instead.

The constructor for IncrementalDocVectors now requires these additional parameters, but it is able to take anything implementing the VectorStore interface rather than being restricted to a TermVectorsFromLucene object. 

BuildPositionalIndex now produces (incremental) docvectors.