Added a test for getMaximumSharedWeight and addition and normalization.
I think getMaxiumumSharedWeight may be broken, or I don't understand it properly. It looks like it is most likely to return zero unless there is an entire row of 1s in the voting record, which seems highly unlikely.

Also uncovered a "last bit gets turned to 1" behavior which I'm not sure is intended.

Some name changing and commenting, let me know if there are any objections or misunderstandings here.