Make the FlagConfig getter methods to be just (e.g.) "dimension()" instead of "getDimension()". - Instead of a dimensionDescription field, we just use the (programmatically accessible) JavaDoc of the public dimension() method.   - This JavaDoc can list the default values used.
We could add a test to the FlagConfigTest class to make sure that
there is programmatically accessible JavaDoc for all public methods.

I think all of these combined will lead to an easy way to document
more clearly and generally "how flags work", and would lead to the
FlagConfig.html page being a very useful "how to" document for what
flags are available.