Working prototype of TTVFL with Lucene 4, and update to IndexFilePositions to create a StandardAnalyzer without any stop words.
Of note, the LuceneUtils AtomicIndex machinery couldn't manage the TASA corpus, which is to say that leaves.get(0) didn't return the entire index. So for testing purposes I used the following kludge that I have not checked in, that I found on the website below:

kludge: this.atomicIndex = this.atomicReader =SlowCompositeReaderWrapper.wrap(compositeReader);
source of something like kludge: http://blog.thetaphi.de/2012/02/is-your-indexreader-atomic-major.html

IncrementalDocVectors doesn't yet work in this context, but I'm able to get some sensible term vectors written out. Performance seems slower than I remember, but some of this may have to do with new Hashtable based implementation, which is easier to read but appears to be presenting some difficulties for the garbage collector. That said, iterating through the terms to generate index vectors also seems to take awhile (perhaps it isn't called BlindinglyFastCompositeIndexWrapper for a reason).