A (GIT repo test and) working first commit of Symmetric Random Indexing, a RI variant we developed awhile back that produces a reduced-dimensional approximation of the term-document matrix multiplied by its transpose.
SRI was first described in: 
Cohen, Trevor, and Roger W. Schvaneveldt. "The trajectory of scientific
discovery: concept co-occurrence and converging semantic distance." Stud
Health Technol Inform 160 (2010): 661-665.

Despite generating a reduced-dimensional approximation of the results of
a matrix multiplication (which is conducted by calculating components of
the scalar product between row vectors individually on a
document-by-document basis), the implementation is straightforward, and
amounts to adding elemental vectors for other terms in a document to the
semantic vector for each term it contains.