First pass at an implementation of skipgram with negative sampling (Mikolov 2013). The main changes are as follows:
(1) The negative sampling algorithm itself, with the scalar product and
vector updates currently implemented within VectorUtils using
netlib-java's blas routines (already a dependency), and sampling of
out-of-context terms in accordance with their frequency^.75.

(2) Subsampling of frequent terms, currently in need of reassessment and
review, as although it does (vastly) speed up training, it does not
improve performance as assorted papers suggest it should. Without
subsampling results on pairwise correlation sets (e.g. wordsim353, MEN)
and Mikolov's analogy set (n=~19,000) match reported results - in the
region of 0.7 correlation and 0.6 accuracy for similarity and analogies
respectively. But with subsampling as implemented currently performance
deteriorates, more so as the threshold is reduced, which should not be
the case according to the literature.

(3) Multiple processing threads (with some initial movements in the
direction of thread safety, though this doesn't yet extend to the
document queue)